errors, ensure configuration is preserved
Error reporting:
When a template function is not defined the error message will include
the serialized node to help users figure out where the problem occurs.

Ensure configuration is present in $model:
Before, each time the configuration was retrieved it was first checked
if it was present.
This check is now moved to the places where users are able to modify
the model and thus may remove it from the map.

- add a new private function templates:process-children
- templates:process will now check for the configuration to be present
- now that the presence of configuration is ensured
  `templates:get-configuration` is no longer necessary as this can be
  expressed with a simple map lookup