Handle <=6.2.0 vs. >= 7.0.0 error response codes
eXist 6.2.0 and earlier issues an error 400 when the fn:error function is called. With 7.0.0-SNAPSHOT (https://github.com/eXist-db/exist/pull/4649), eXist issues an error 500. The precise error code is outside the scope of this package - it’s an issue decided by the eXist API - so we can accept either here.